Добавляет задание 7.5. Стой! Кто идёт?!
