+++ controllers/TasksController.php
@@ -3,30 +3,12 @@ namespace app\controllers;
 
 use app\models\Task;
 use app\models\TaskFilterForm;
-use app\models\User;
 use yii\data\ActiveDataProvider;
 use yii\db\Expression;
-use yii\filters\AccessControl;
-use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 
-class TasksController extends Controller
+class TasksController extends SecuredController
 {
-    public function behaviors()
-    {
-        return [
-            'access' => [
-                'class' => AccessControl::class,
-                'rules' => [
-                    [
-                        'allow' => true,
-                        'actions' => ['index', 'view' ],
-                        'roles' => ['@'],
-                    ],
-                ],
-            ],
-        ];
-    }
 
     public function actionIndex()
     {
